Remove unnecessary DotNetAdapterSDK.csproj
In 757b75803b30a61fa91f6ae4aa45503788ef984c we made a csproj that will generate a NuGet Package. This makes the DotNetApapterSDK.csproj and Sample/AdapterLab.sln redundant. Accordingly we needed to change the references to the NUnit test and AdapterLab projects and the DotNetAdapterSDK.sln to reference the new project.

NuGet package re-targeted to 4.5. Bumped the NUnit and AdapterLab project frameworks to 4.5 to be able to use mtconnect_adapter_sdk.

Changed the project file for AdapterLab to use NetFramework4.5 format.

Replaced NUnit dlls with NuGet references in the NUnitTestsProject